Factor out WebView fragment from ContentController.
It's complex, generic and has no interdependencies into
ContentController.
